+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 3.3.23 - Unreleased
 ### Changed
 * Revert "Social image transforms for remote volumes no longer default to not generating the image transforms immediately (they fall back on whatever your system setting is)" because it caused slowdown when working with remote volumes
+* Ensure the URLs in the breadcrumbs JSON-LD adheres to the `addTrailingSlashesToURls` general config setting
 
 ### Fixed
 * Use unicode safe string functions to truncate meta values > 4096 characters before parsing them