+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 * Revert "Social image transforms for remote volumes no longer default to not generating the image transforms immediately (they fall back on whatever your system setting is)" because it caused slowdown when working with remote volumes
 * Ensure the URLs in the breadcrumbs JSON-LD adheres to the `addTrailingSlashesToURls` general config setting
+* Added better checks for `InStock` for Product JSON-LD schema
 
 ### Fixed
 * Fixed an issue where SEOmatic could permanently change "robots" to "none" inadvertently in edge-case circumstances